models/res_net_50/requirements.txt
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.19.0
matplotlib>=3.3.0
tqdm>=4.49.0
pillow>=8.0.0
scikit-learn>=0.23.0
seaborn>=0.11.0 